added send/receive ttft as well as added to validation table

+++ create_layer_files.py
@@ -10,7 +10,7 @@ from config import (
 os.makedirs(LAYERS_DIR, exist_ok=True)
 
 print("Loading full model...")
-model = AutoModelForCausalLM.from_pretrained(MODEL_PATH)
+model = AutoModelForCausalLM.from_pretrained(MODEL_PATH, dtype=torch.float16)
 full_state = model.state_dict()
 
 # Save each layer separately
